set up front page buttons
